abstract out the common functions, add a py3.8 transform as well.
